add order history

+++ api/fixtures/products.js
@@ -2,6 +2,6 @@ const products = require('./data/products.json')
 
 module.exports = products.map((product) => ({
   ...product,
-  price: parseInt(product.price),
+  price: parseFloat(product.price),
   description: product.description.replace(/(<([^>]+)>)/gi, '')
 }))
